# openapi.yaml
openapi: 3.0.3
info:
  title: NetVisor Suite API
  version: 1.0.0
  description: REST API for documenting, administering, and monitoring office network infrastructure.
servers:
  - url: /api/v1
paths:
...
  /vlans:
    get:
      tags: [Network]
      summary: List VLANs
      parameters:
        - in: query
          name: limit
          schema: { type: integer, maximum: 500 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        '200': { description: Array of VLANs }
    post:
      tags: [Network]
      summary: Create VLAN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vlan_id, name]
              properties:
                vlan_id: { type: integer, minimum: 1, maximum: 4094 }
                name: { type: string, maxLength: 80 }
                subnet: { type: string, maxLength: 64 }
                gateway: { type: string, maxLength: 64 }
                description: { type: string, maxLength: 2000 }
      responses:
        '200': { description: Created VLAN }
        '400': { description: Validation error }
  /vlans/{id}:
    patch:
      tags: [Network]
      summary: Update VLAN
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vlan_id: { type: integer, minimum: 1, maximum: 4094 }
                name: { type: string, maxLength: 80 }
                subnet: { type: string, maxLength: 64 }
                gateway: { type: string, maxLength: 64 }
                description: { type: string, maxLength: 2000 }
      responses:
        '200': { description: Updated VLAN }
        '404': { description: VLAN not found }
    delete:
      tags: [Network]
      summary: Delete VLAN
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted, devices detached }
        '404': { description: VLAN not found }
  /devices:
    get:
      tags: [Network]
      summary: List devices
      parameters:
        - in: query
          name: limit
          schema: { type: integer, maximum: 500 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        '200': { description: Array of devices }
    post:
      tags: [Network]
      summary: Create device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 80 }
                ip: { type: string, format: ipv4 }
                device_type: { type: string, enum: [Router, Switch, Server, Workstation] }
                vlan_id: { type: integer }
                mac: { type: string, maxLength: 32 }
                location: { type: string, maxLength: 120 }
                monitored: { type: boolean }
      responses:
        '200': { description: Created device }
        '400': { description: Validation error }
  /devices/{id}:
    patch:
      tags: [Network]
      summary: Update device
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80 }
                ip: { type: string, format: ipv4 }
                device_type: { type: string, enum: [Router, Switch, Server, Workstation] }
                vlan_id: { type: integer }
                mac: { type: string, maxLength: 32 }
                location: { type: string, maxLength: 120 }
                monitored: { type: boolean }
      responses:
        '200': { description: Updated device }
        '404': { description: Device not found }
    delete:
      tags: [Network]
      summary: Delete device
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted }
        '404': { description: Device not found }
...
  /monitor/check/{id}:
    post:
      tags: [Monitoring]
      summary: Ping a single device
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Ping result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  name: { type: string }
                  ip: { type: string }
                  status: { type: string, enum: [up, down] }
                  rttMs: { type: number, nullable: true }
        '404': { description: Device not found or missing IP }
...
  /monitor/history/{id}:
    get:
      tags: [Monitoring]
      summary: Last 50 checks for a device
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Array of history entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer }
                    device_id: { type: integer }
                    status: { type: string, enum: [up, down] }
                    rtt_ms: { type: number, nullable: true }
                    checked_at: { type: string }
  /issues:
    get:
      tags: [Incidents]
      summary: List issues
      parameters:
        - in: query
          name: limit
          schema: { type: integer, maximum: 500 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        '200': { description: Array of issues }
    post:
      tags: [Incidents]
      summary: Create issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 140 }
                description: { type: string, maxLength: 2000 }
                severity: { type: string, enum: [Low, Medium, High], default: Medium }
                status: { type: string, enum: [Open, In Progress, Resolved, Closed], default: Open }
                device_id: { type: integer }
                outlet_id: { type: integer }
                reporter: { type: string, maxLength: 80 }
      responses:
        '200': { description: Created issue }
        '400': { description: Validation error }
  /issues/{id}:
    patch:
      tags: [Incidents]
      summary: Update issue
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 140 }
                description: { type: string, maxLength: 2000 }
                severity: { type: string, enum: [Low, Medium, High] }
                status: { type: string, enum: [Open, In Progress, Resolved, Closed] }
                device_id: { type: integer }
                outlet_id: { type: integer }
                reporter: { type: string, maxLength: 80 }
      responses:
        '200': { description: Updated issue }
        '404': { description: Issue not found }
    delete:
      tags: [Incidents]
      summary: Delete issue
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted }
        '404': { description: Issue not found }
...